+++ Sources/StreamChat/APIClient/APIClient.swift
@@ -4,6 +4,16 @@
 
 import Foundation
 
+/// The retry policy to use by the API retry mechanism
+struct RetryOptions {
+    /// How many attempts should be performed at most
+    let maxRetries: Int = 3
+    /// How much time there should be in between attempts, this function is usually dependant on the amount of attempts already performed
+    let backoff: (Int) -> Double = {
+        min(max(Double($0) * Double($0), 0.5), 6.0)
+    }
+}
+
 /// An object allowing making request to Stream Chat servers.
 class APIClient {
     struct RequestsQueueItem {
@@ -28,6 +38,9 @@ class APIClient {
     /// Used for synchronizing access to requestsQueue
     private let requestsAccessQueue = DispatchQueue(label: "io.getstream.requests")
     
+    /// Used for request retries
+    private let requestsRetriesQueue = DispatchQueue(label: "io.getstream.request-retries")
+
     /// Stores request failed with token expired error for retrying them later
     private var requestsQueue = [RequestsQueueItem]()
     
@@ -62,6 +75,63 @@ class APIClient {
     
     deinit {
         requestsQueue = []
+    }
+
+    /// Performs a network request and retries in case of network failures
+    ///
+    /// - Parameters:
+    ///   - endpoint: The `Endpoint` used to create the network request.
+    ///   - timeout: The timeout in seconds for the API request.
+    ///   - retryOptions: The `RetryOptions` used to retry the request.
+    ///   - completion: Called when the networking request is finished.
+    func request<Response: Decodable>(
+        endpoint: Endpoint<Response>,
+        timeout: TimeInterval = 60,
+        retryOptions: RetryOptions,
+        completion: @escaping (Result<Response, Error>) -> Void
+    ) {
+        requestsRetriesQueue.async { [weak self] in
+            self?.request(attempt: 0, endpoint: endpoint, timeout: timeout, retryOptions: retryOptions, completion: completion)
+        }
+    }
+
+    private func request<Response: Decodable>(
+        attempt: Int,
+        endpoint: Endpoint<Response>,
+        timeout: TimeInterval,
+        retryOptions: RetryOptions,
+        completion: @escaping (Result<Response, Error>) -> Void
+    ) {
+        request(endpoint: endpoint) {
+            let backoff = retryOptions.backoff(attempt)
+
+            guard case let .failure(error) = $0 else {
+                return completion($0)
+            }
+
+            // we only retry transient errors like connectivity stuff or HTTP 5xx errors
+            guard ClientError.isEphemeral(error: error) else {
+                return completion($0)
+            }
+
+            let offlineErrorCodes: Set<Int> = [NSURLErrorDataNotAllowed, NSURLErrorNotConnectedToInternet]
+            let connectionError = offlineErrorCodes.contains((error as NSError).code)
+
+            // give up after `retryOptions.MaxRetries` unless its a connection problem
+            if attempt <= retryOptions.maxRetries && !connectionError {
+                return completion($0)
+            }
+
+            self.requestsRetriesQueue.asyncAfter(deadline: .now() + backoff) {
+                self.request(
+                    attempt: attempt + 1,
+                    endpoint: endpoint,
+                    timeout: timeout,
+                    retryOptions: retryOptions,
+                    completion: completion
+                )
+            }
+        }
     }
     
     /// Performs a network request.